stacks/python/requirements.txt
# OpenTelemetry SDK and OTLP exporter for SigNoz (starter)
# Update: pip install -U opentelemetry-api opentelemetry-sdk opentelemetry-exporter-otlp-proto-grpc opentelemetry-instrumentation

opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp-proto-grpc>=1.28.0
opentelemetry-instrumentation>=0.49.0
opentelemetry-instrumentation-wsgi>=0.49.0
opentelemetry-exporter-otlp-proto-grpc-metrics>=1.28.0
